+++ Battle_interface/land/battle_mancommand.cpp
@@ -58,7 +58,7 @@ long BIManCommandList::CommandAdding()
 	if( !pAttr ) return 0;
 	long attrQuant = pAttr->GetAttributesNum();
 
-	for(long i=0; i<attrQuant; i++)
+	for(size_t i=0; i<attrQuant; i++)
 	{
 		ATTRIBUTES * pA = pAttr->GetAttributeClass(i);
 		if(pA== nullptr) continue; // нет такого атрибута
@@ -81,7 +81,7 @@ long BIManCommandList::UserIconsAdding()
 	if( !pAttr ) return 0;
 	long attrQuant = pAttr->GetAttributesNum();
 
-	for(long i=0; i<attrQuant; i++)
+	for(size_t i=0; i<attrQuant; i++)
 	{
 		ATTRIBUTES * pA = pAttr->GetAttributeClass(i);
 		if(pA== nullptr) continue; // нет такого атрибута
@@ -104,7 +104,7 @@ long BIManCommandList::AbilityAdding()
 	if( !pAttr ) return 0;
 	long attrQuant = pAttr->GetAttributesNum();
 
-	for(long i=0; i<attrQuant; i++)
+	for(size_t i=0; i<attrQuant; i++)
 	{
 		ATTRIBUTES * pA = pAttr->GetAttributeClass(i);
 		if(pA== nullptr) continue; // нет такого атрибута
